+++ src/asp/PcAlign/pc_align_utils.h
@@ -106,8 +106,14 @@ void calc_extended_lonlat_bbox(vw::cartography::GeoReference const& geo,
 /// Compute the mean value of an std::vector out to a length
 double calc_mean(std::vector<double> const& errs, int len);
 
-/// Compute the standard deviation of an std::vector out to a length
-double calc_stddev(std::vector<double> const& errs, double mean);
+/// Compute the standard deviation of an std::vector
+double calc_stddev(std::vector<double> const& errs);
+
+/// Compute the mean absolute error of an std::vector
+double calc_mae(std::vector<double> const& errs);
+
+/// Compute the root mean square error of an std::vector
+double calc_rmse(std::vector<double> const& errs);
 
 /// Calculate translation vector between the centers two point clouds
 void calc_translation_vec(Eigen::MatrixXd const& initT,